feat(refactor): Basehandle
