+++ trailblazer/store/core.py
@@ -1,7 +1,10 @@
+import alchy
+
 from trailblazer.store.crud.create import CreateHandler
 from trailblazer.store.crud.delete import DeleteHandler
 from trailblazer.store.crud.read import ReadHandler
 from trailblazer.store.crud.update import UpdateHandler
+from trailblazer.store.models import Model
 
 
 class CoreHandler(
@@ -14,3 +17,8 @@ class CoreHandler(
 
     def __init__(self):
         pass
+
+
+class Store(alchy.Manager, CoreHandler):
+    def __init__(self, uri: str):
+        super(Store, self).__init__(config=dict(SQLALCHEMY_DATABASE_URI=uri), Model=Model)
